Parallelize loading product workflow
